requirements.txt
annotated-types==0.7.0
anthropic==0.55.0
anyio==4.9.0
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.6.15
click==8.2.1
distro==1.9.0
Flask==3.1.1
flask-cors==6.0.0
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
lxml==5.4.0
MarkupSafe==3.0.2
# openai==1.88.0  # Removed - migrated to Anthropic
# psycopg2-binary>=2.9.5
# psycopg[binary]>=3.1.0
# PostgreSQL drivers temporarily disabled due to compatibility issues
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
PyPDF2==3.0.1
python-docx==1.2.0
python-dotenv==1.1.0
python-magic==0.4.27
sniffio==1.3.1
SQLAlchemy==2.0.41
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
Werkzeug==3.1.3
bleach==6.1.0

# Async processing dependencies
celery==5.4.0
redis==5.2.0
kombu==5.4.2

# Testing dependencies
pytest==8.3.0
pytest-cov==5.0.0
pytest-flask==1.3.0
pytest-timeout==2.2.0
reportlab==4.2.5  # For PDF generation in tests

# Additional dependencies for async tasks
pandas==2.2.3  # For data export
openpyxl==3.1.5  # For Excel export
Pillow==11.0.0  # For image processing
